Fix build issue with MacOS

+++ libs/vm/include/vm/array.hpp
@@ -90,7 +90,7 @@ struct Array : public IArray
       return {};
     }
 
-    auto element = GetIndexedValue(AnyInteger(elements.size() - 1u, TypeIds::Int32))
+    auto element = GetIndexedValue(AnyInteger(static_cast<uint64_t>(elements.size() - 1u), TypeIds::Int32))
                        .template Move<TemplateParameter>();
 
     elements.pop_back();
